EventTableViewCell.m: give service name a little more space on iphone
this should be a reasonable size, you should be able to identify the
service name while still having a fair distance from the timeframe.
given that you should be able to easily find out which starts/ends where
the distance should be fine, though it could be a little bigger imo.