properly force config list into foreground when host unreachable
well, not completely yet but much better. all the tabs and the other
list (root) are guarded so the notification is only missing when one is
"deep" in any view but this is not different to the behaviour of 0.3.0.

also fixes the feature checked for media player on ipad (it's NOT
recordings ;-))