add code to start search on imdb (disabled though)
the url handler of imdb does not seem to be made for full titles but
actual keywords. right now there are two choices: we either correctly
urlencode the string which results in imdb being unable to find it or
we don't and ios disallows opening the url.
as neither appears to be a good idea just leave the code in place for
future use.
this is still up for some further testing. we seem to be able to work
around this issue partially by not doing urlencode but rather encode for
a post request which imdb seems to do internally (e.g. turn "the
simpsons" into "the+simpsons") and subsequently get results, but as imdb
still shows the string wrong in its gui this also does not seem to be
intended.