MultiEPGListController: manage "now" timestamp for cells
includes a timer which refreshes the timestamp every five minutes if it
was previously visible and the view is shown. we might want to consider
using a shorter interval or to drop the timer at all though. depends on
how well the devices handle a frequent redraw. but for now this is fine.