improve "now" handling in multi epg
instead of using a "complex" nsdate object for the timestamp just use
the "simple" nstimeinterval which is the only thing we're interested in
anyway. -1 used as illegal value, we should only get values in the range
of (0,7200) anyway.

also changes the timer to use 60s as interval and to be always active
when the view is shown.