ConfigListController.m: allow selection while editing
first step to change behavior of config list. in a future commit
selecting a row while not editing will connect to said host, and the
config view will only open if editing is enabled.
this will significantly speed up changing the active connection and
existing users should be able to either catch up on this change by
themselves or by reading the release notes :D