add welcome screen on first launch
it's also presented to users of an earlier version in case they were
unable to configure it properly before. also added (though of no use in
this version) is a changelog view on upgrades. it can be used to remind
users of important changes or just point them at fixed bugs. i kinda
feel no one but me reads the changelogs unless they are forced to :-)