NeutrinoConnector.m: use sref for timers if we know it
try harder to use sref when communicating with timerd interface as they
are easier to handle. check for invalid sref should be superfluous in
addTimer: but lets keep it there for now.
we should figure out a way to get sref for all services though but using
format=id when requesting timers is no viable alternative since then
we'd have to find a way to get sname for all services since we don't
want to tell the user the record is scheduled on e.g. 2718f001d175.