TouchXML: update sources snapshot
also adds two modifications to origin which avoid putting nil in
cxmlnodes children array or xpath results. internally the code does not
handle this and therefore crashes in those cases. since i don't know the
effects of this (and if it actually fixes the problems i was reported)
keep this modification local.