fix "playing" view for neutrino and enable it
this is a little hackish though, as we retrieve the 50 first events and
just hope, that the current ones are in that list. so it's possible that
we retrieve way too many events (at least with the default settings of
keeping only events in the last hour in the cache) or too few (e.g. if
you keep a whole day on a busy station). another possible speedup would
be to use the epg cache in the full version, but this might lead to
false (invalid cache because of changed station schedule) or no
(outdated cache) results.
should consider that for an upcoming version though, as checking locally
for results first is way faster than retrieving a possibly long and
useless list from a remote host and we don't care about an outdated
cache in the multi epg either.