maybe fix the CFHash crash
just guessing here, but it appear the crash occurs when trying to remove
or add credentials from the cache. so save them for this session and
additionally make sure username and password do not get released by
another thread.
also makes sure that there are actually any credentials (non-empty
username), otherwise we abort the request earlier.