MediaPlayerController.m: fix possible crash, immediately hide sheet if reshown
this fixes the sheet being shown when re-selecting the media player from
the tab bar (actually I think it is shown, but just hidden again before
ever being visible).
this also automatically hides the sheet if coming back from background,
though it's briefly visible in that case. we might want to considere
listening to the notification and hiding it explicitely when in
background.