run view specific operations into a NSOperationQueue
instead of creating individual threads, uses one queue which is kept
inside the RemoteConnectorObject singleton. this does a) appear to make
things a little faster (as the threads are already pre-created by the
queue) and b) allow us to eventually empty the queue when switching
views, which can save traffic and cpu time.