MultiEPGTableViewCell.m: try to fix group 414
i assume this is a race condition, but work around it as good as we can.
another approach would be to translate the touched point to a point in
time and check if it lies within the bounds of this event, but that
would behave badly on e.g. sky select.