MultiEPGTableViewCell.m: some fixes
adds prepareForReuse which is supposed to fix some crashes that i worked
around in a very ugly way (try/catch on array out-of-bounds read ;)),
add another line which might have resulted in a crash to the same
workaround (that one still crashed once - and the first time ever - for
me after prepareForResuse was added), improve debug output in the
workaround code.
also fixes path (event borders) being disrupted if event was not at
left/right border of timeframe.