PackageManagerListController: improve commit a bit
now commit takes you to a list with just the selected packages and if
you press commit again the changes will be applied. toolbar changes to
just abort/commit and deselection is still possible (though no
reselection as the items disappear upon deselection).
also adds a warning if no packages are selected and commit is pressed.