MultiEPGTableViewCell.m: rework current event detection algorithm
now properly works with events that began on a previous page or if they
are the only event on that page.
a little overly complex (we might not want to use the "precalculated"
lines in the future but instead generate them on the fly - can't be
much worse performance-wise than now anyway) but does its job.