ServiceListController.m: disable now/next when editing
otherwise modified service names are not shown. we might want to
put some code in the webif to change this as this still affects our
regular service list.
but given i retried to change a servicename several times and did not
get what failed we might want to keep it like this for now :D