ServiceListController.m: improve button handling
also fixes radio mode mixup,
hides radio mode if a delegate.
on the iphone if editing is possible the edit button is the right bar
button item and multi epg moves to the right in a toolbar we add in this
case.
nothing changes on the ipad or if editing is not possible.