LocationListController: allow to add/remove locations
this also moves the cancel button to the left, as the normal spot for
the editbutton is on the right.
we might want to limit the button to the ipad now though, or maybe put
two buttons on the right there (which would not leave much space for the
title though).