ServiceListController: allow to force "simple" list
showNowNext is a new externally visible property replacing the old
private supportsNowNext. behavior is the same, but this allows us to
force-disable now/next - we needs this for 'All Services' which
are already pretty heavy on the device but with the additional
information this pushes my ipad to the limit and i only have two sources
of services. :P