ServiceListController: basic multi selection
hold service for 1s to select it, last service to be selected with a
regular tap - allows for add to bouquet and cancel selection.

also added is search capability for the service list if in app is bought
(glue code not yet present) or in the full version.

there are still some rough edges, but overall this is pretty nice :D