BouquetListController.m: allow to add a new bouquet if delegate for slist
if editing and delegate for a ServiceListController we now allow to add
a new bouquet. also properly sets editing mode now if editing was
initiated before the table view was instantiated.

this is a little flawed though since the "original" bouquet list is NOT
reloaded, hence the user has to reload it manually.
still have to figure out how to handle this best :)