exception to background picon loading if image was loaded before
if the image was loaded before we don't need to recalculate the name or
get delayed by processing it, so in that case just handle things
internally.

this improves the user experience a lot because the picons no longer
delay the loading process but still appear immediately if they were
shown before (in the current bouquet).