BouquetListController.m: make one debug exception a nslog
previously it was very unlikely to be triggered, but now in editing mode
the list does have one element so one can expect the user to tap the
respective row. though this was never a crash for the user but just in
debug builds this is reason enough to just make it a warning instead of
a failure.