ServiceListController.m: mostly reformatting
was trying an approach for loading the picons more closely to the lazy
table view sample code, but it was way to slow for my taste. hence this
is the disabled code to enable this and some basic internal restructuring
which was overdue. could still need more of it though :D