better fix for possible crash with tone interruption listener
if the view disappears any possible way we stop playing audio and the
client data in the callback could point to an invalid reference if the
signal finder was deallocated due to memory constraints.
we now initialize the audio session in the app delegate and use it as
target for the (currently dummy) interruption callback.