MKStoreKit: initial checkin
uses custom fork in case we need to fix anything quick and the current
"ARC" branch which contains some further fixes to arc support.

to make this work properly, drop our local copy of NSData+Base64 (as it
comes with MKStoreKit and their copy contains some fixes) and don't
include SFHFKeychainUtils from MKStoreKit as it is already build with
ShareKit.