DreamoteConfiguration.m: theme change improvements
by forcing a white background color instead of nil in the default theme
we fix black labels when changing back from the black background color,
probably nil was understood as "invalid" or "keep current color".

fun fact: with the previous approach the only item in the 'other list'
working correctly on the ipad was the currently selected (settings) one
after the change - now this one is broken for some reason. but only when
directly switching from night to default, not when changing to blue in
between...