multi selection in movie list and base cell on MultiSelectionTableViewCell
multi selection when not in editing mode will bring up the delete
confirmation as before, but when editing multi selection is the way to
go. deletion is done in sequence (not in parallel) for performance
reason - progress hud keeps track of the progress.

this also modifies the movie objects so you can cache the formatted date
in them, as we would otherwise need to recalculate it on each draw
operation which might slow us down.