SimpleMultiSelectionListController: initial checkin
similar to SimpleSingleSelectionListController this is a reusable
component which allows to select any number of items from a list of
one or more items.
this is a more generic approach to presenting the tag list and required
for the movie list. hope i can reuse this for the list of timer tags
though :)