MovieViewController.m: allow to move recordings
if the backend supports it we add another button to the bottom of the
list (Move). upon selection we show the location list and subsequently
instruct the stb to move the recording.
smaller ui stuff left (error if location or moving not supported for
some reason and delete the recording from our source list if moved
successfully).