NeutrinoConnector.m: try to work around bogus streaming url
some versions of neutrino hd seem to return a double port in the
streaming url (instead of just the hostname it uses the hostname:port
combination). try to work around this by replacing hostname:port by just
hostname in the returned string. the external players don't seem to like
it :D