AutoTimerViewController: untested support for new search type selection
uses a simple single selection list and shares the code with avoid
duplicate description.
i merged the two paths because i felt this was easier to read in the end
because after all except for basically three lines of code the same
stuff happens in both paths - so the two ifs are easier to understand
than trying to spot the difference in two pretty much equal blocks of
code.