OtherListController.m: some theoretical code to transfer split view reference
needed for an upcoming commit. if the master view expects to know about
its split view we will now tell it which one is actually its current
presenter (and change it back upon return). this allows for some crazy
stuff where the pushed master view changes the split view (and
possibly unravel the very fabric of the space-time continuum and destroy
the entire universe! - wait, I think I heard this one before :D).