make use of NSObject+Queue (and initial checkin of it)
this allows us to reduce the load on the main thread a bit by calling it
less quickly.

currently we make use of it only on the "high profile" views (services,
events, movies) and their respective sibling views (bouquets,
locations). without the latter the effect would be less noticeable
because the thread would still be overloaded, though for less time.

we could still optimize this a bit by e.g. only loading off to the main
thread if we actually can dispatch multiple items instead of always
going there (sometimes just to check this very fact), but this is still
a work in progress :D