make initial load of service list faster on the iphone
the previous code regarding picon loading was flawed in the full version
as requests were forwarded to the multi epg which would load the picon
on the main thread (*ouch*) - this lead to serious lag on the iphone,
even in the regular service list.
now we a) don't load picons in the multi epg if it isn't shown and b)
implement the 'batch dispatch' code to give things a little more speed.

on top of that, we also don't animate adding rows in the regular service
list (not now/next) if we received more than kBatchDispatchItemsCount
(10) items at once - this should not happen any longer but it was
possible in temporary test code which eventually might be added again.