BouquetListController.m: don't try to load picons
while it is theoretically possible that a user has picons for her
bouquets it is a) unlikely and b) not officially supported.
if we want to support them, we should add a "background loading"
mechanism just as the service list has instead of trying to load them
all on the main thread as the cells are drawn.