GenericService: also require any sref for a valid service
now we can just toggle the flag to mark a service invalid, but to be
recognized as "valid" we need both this flag set to yes AND anything as
service reference.
this fixes the cases where a known invalid service (e.g. Error
retrieving Data) was recognized as valid since the detection behavior
was changed.