OtherListController.m: try to fix possible crashes
force reloadData to happen on the main thread and don't do more work
than needed in the reconnect handler.
also tries to avoid possible race condition in cellForRowAtIndexPath if
an old reload is still happening (items are loaded) but the list array
is already being modified from another reconnect (or - in the case i ran
into - a finished version/feature detection).