Enigma2Connector: some feature detection enhancements
detection runs the first time the version is checked (not just on our
own background thread), this fixes possible gui confusions (item popping
up briefly before we detect it's not actually supported).
also adds detection for webif 1.5b4 which added external support and use
that as minimum version to run the additional feature check on. uses
external xml file on 1.7.0 to increase speed a bit. a little more data
but one instead of three requests and without possible side effects.

also adds check for bouquet editor (though it is part of the dmm
standard distribution this might not be true for other images,
especially older ones).