ConfigListController.m: some clarifications
if the current connection does not support the service editor the user
is hinted to install the appropriate plugin in enigma2. in theory we
could also show messages specific to other backends if we are currently
connected to one, but that's up to future me to implement.

also unifies some localizable strings.