create new neutrino specific service
similar to the approach in enigma1 - based on GenericService but extends
it by the piconName property (which was removed from the base class).
following this change adjust NeutrinoServiceXMLReader to use the new
service object and remove the neutrino stuff from the generic picon
loading code and overwrite the whole method in NeutrinoService.
this might optimize things a tiny bit, but first and foremost this
improves readbility and makes the code easier to understand.