Enigma2Connector.m: some "OpenWebif" handling
force the version to 1.5beta3 and if possible print a warning - it
contains bugs, is not even close to being feature complete and I still
find at least the name strongly offensive - not to forget that it is
anything but open and contradicts its own philosophy by their license
which forbids use on a dreambox running the original software.

well, I'm going to comply with their wishes, so this is absolutely
untested and the warning is going to stay.