rework accessibility handling
instead of assigning the properties (which copies the strings so I
consider it bloat in most cases - when accessibility is not required) we
overwrite the getter method which subsequently allows us to do more
complicated transformations there.
this also changes the timer cell to read the time if not editing (so its
title/time if not editing, title/selected status if editing).