FileListView.m: try to avoid possible race condition
sometimes in selectPlayingByTitle there still is a crash due to problems
with the table view. Given the table is no longer modified if _reloading
is false I assume it was not false "long enough", so move this around a
bit.
It could however also be an odd-timed reload but I have to take a closer
look at this one.