SimpleSingleSelectionListController.m: fix double-callback
as all the other times, unset the instance reference to the callback
before calling it to prevent possible race conditions.
removing the code where we'd normally unset the callback afterwards and
just re-save the callback in the other code path.

this probably copies it again but better than weird crashes ;)