try to increase lifetime of errors
this might have been specific to svdrp though (and is possibly solved
now), but just in case make sure the argument to the helper function is
retained and only that to setArgument is unretained. since it retains
the argument for us (or at least is supposed to) this SHOULD take care
of the problem.
a simple cast instead of a helper variable might be easier, but casting
has gotten hard with arc and I just don't know which is the right one
currently ;)