BouquetListController.m: fix mepg button not being shown on the iphone
the service list would show the toolbar before we are able to hide it,
as a result the toolbar would be hidden given the user purchased the
service editor when shown from the bouquet list (rotation or re-entering
the app would show it though).
i hope this breaks nothing else, but besides introducing more "inside
knowlege" into the bouquet list this should be fine.