update mbprogresshud and enable arc
since we cannot disable arc for the header and I'm scared b/c of
possible side effects of using arc in the header but no arc in the
actual module just use arc.
this is however incomplete as upstream still assumes the weak keyword to
be available which is NOT necessarily the case - but I assume this will
be fixed in a future pull (I actually thought a respective pull request
was already merged :D).