OtherListController.m: try harder to get view controller on ipad
somehow the visible view controller seems to have been nil once in a
crash report, so just to be sure also try to get the first object from
the view controllers array and avoid the crash if this fails too.

not really sure if this fixes the issue and this might even confuse the
users (since I did not add any feedback in case of an error) but I can
take a closer look at this again later.