AutoTimerViewController.m: code style fix
instead of manually generating the index path two times just use the
input path (which is correct anyway).
also replaces to _tableView instances with tv which SHOULD point to the
same object but _tableView also forces a reference to self and since we
also use the local indexPath object there is no reason not to do this.