{Event,Movie}ViewController.m: show shk action sheet from tab bar
for some reason cancel worked from the event view but not the movie view
(though they used pretty much the same code).
change both instances from showInView to showFromTabBar which the app
delegates tab bar controller's tab bar instance which SHOULD fix this
for the movie view and not change the behavior in the event view.