add class for asynchronous network requests
blocks are used as callbacks, and base class was slightly restructured
to make the asynchronous version easier to write. the synchronous class
method now calls an internal (synchronous) method to do the request and
our asynchrounous version just calls it from a block running in
background.
no idea if this actually works as expected and it is currently not
included in the binary - eventually we want to use it for the package
manager (commit changes / upgrade) as the user currently has no idea
what is going on there.