one step in the direction of live multi epg with e2
add the necessary code to allow requesting the epg of a bouquet in a
given interval - it is however not used anywhere so far. not sure what
the optimized place to implement it is.
probably in the multi epg itself if the epgcache request did not return
any events (and in return fill the information back to the epgcache) -
but it feels kinda fishy and overly complicated.